chore: reduce deprecated code from migration
