+++ src/main/java/bflow/auth/services/UserService.java
@@ -19,21 +19,6 @@ public interface UserService {
     User findById(UUID id);
 
     /**
-     * Resolves an OAuth2 user by checking email and provider consistency.
-     * @param email user email.
-     * @param providerId external provider ID.
-     * @param provider the provider (e.g., GOOGLE).
-     * @param emailVerified whether the email is verified.
-     * @return the resolved User entity.
-     */
-    User resolveOAuth2User(
-            String email,
-            String providerId,
-            AuthProvider provider,
-            boolean emailVerified
-    );
-
-    /**
      * Retrieves the user profile for a given user ID.
      * @param userId the unique identifier of the user.
      * @return the user profile response.
